Improve daily budget precision and menu bar usage display
- dailyWeeklyBudget now returns Double instead of Int (e.g. 14.17 vs 14)
- Menu bar always shows today's usage as 0-100% of daily budget
- Panel shows precise budget as ~14.2%/day
- Tooltip updated to show today's usage percentage of budget

+++ Sources/UsageData.swift
@@ -93,10 +93,10 @@ struct UsageResponse: Codable {
     }
 
     /// Weekly % remaining divided by days remaining — how much weekly budget per day.
-    var dailyWeeklyBudget: Int? {
+    var dailyWeeklyBudget: Double? {
         guard let days = daysUntilReset else { return nil }
         let remaining = Double(max(0, 100 - (sevenDay?.utilization ?? 0)))
-        return Int(remaining / Double(days))
+        return remaining / Double(days)
     }
 }
